haha

+++ app/src/main/java/com/dong/soufang/exception/CrashHandler.java
@@ -2,8 +2,11 @@ package com.dong.soufang.exception;
 
 import android.content.Context;
 import android.content.Intent;
+import android.os.Looper;
 import android.util.Log;
+import android.widget.Toast;
 
+import com.dong.soufang.MainApplication;
 import com.dong.soufang.activity.WelcomeActivity;
 import com.dong.soufang.util.LogUtils;
 
@@ -37,11 +40,32 @@ public class CrashHandler implements Thread.UncaughtExceptionHandler {
 
     @Override
     public void uncaughtException(Thread thread, Throwable ex) {
-        LogUtils.e(Log.getStackTraceString(ex));
-        Intent intent = new Intent(mContext, WelcomeActivity.class);
-        intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-        mContext.startActivity(intent);
-        android.os.Process.killProcess(android.os.Process.myPid());
+        if (!handleException(ex) && mDefaultHandler != null) {
+            mDefaultHandler.uncaughtException(thread, ex);
+        } else {
+            LogUtils.e(Log.getStackTraceString(ex));
+            Intent intent = new Intent(mContext, WelcomeActivity.class);
+            intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+            mContext.startActivity(intent);
+            MainApplication.getInstance().exitApplication();
+            android.os.Process.killProcess(android.os.Process.myPid());
+        }
+    }
 
+    private boolean handleException(Throwable ex) {
+        if (ex == null) {
+            return false;
+        }
+
+        new Thread(new Runnable() {
+            @Override
+            public void run() {
+                Looper.prepare();
+                Toast.makeText(mContext, "很抱歉，程序出现异常，即将退出", Toast.LENGTH_SHORT).show();
+                Looper.loop();
+            }
+        }).start();
+
+        return true;
     }
 }
